Calculate K-factor

Calculate K-factor

Usage

k_factor(sand, silt, clay, soc, method = c("williams1983"))

Arguments

sand

SpatRaster. Single layer raster with sand proportion (%). Outcome of the aggregate_soilgrids()

silt

SpatRaster. Single layer raster with silt proportion (%). Outcome of the aggregate_soilgrids()

clay

SpatRaster. Single layer raster with clay proportion (%). Outcome of the aggregate_soilgrids()

soc

SpatRaster. Single layer raster with weight percentage of organic carbon (%). Outcome of the aggregate_soilgrids()

method

string. One or more of these options: 'williams1983' (see Details)

Value

SpatRaster

Details

Currently, this function supports the following equations:

'williams1983' -- An equation from Williams and Renard (1983) as cited in Chen et al. (2011)

References

Williams, J. R., K. G. Renard, and P. T. Dyke. “EPIC: A New Method for Assessing Erosion’s Effect on Soil Productivity.” Journal of Soil and Water Conservation 38, no. 5 (September 1, 1983): 381–83.

Chen, Liangang, Xin Qian, and Yong Shi. “Critical Area Identification of Potential Soil Loss in a Typical Watershed of the Three Gorges Reservoir Region.” Water Resources Management 25, no. 13 (June 25, 2011): 3445. https://doi.org/10.1007/s11269-011-9864-4.

Examples

library(purrr)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

sand <- get_soilgrids(v, layer = "sand")
#> Downloading Sand rasters...
silt <- get_soilgrids(v, layer = "silt")
#> Downloading Silt rasters...
clay <- get_soilgrids(v, layer = "clay")
#> Downloading Clay rasters...
soc <- get_soilgrids(v, layer = "soc")
#> Downloading SOC rasters...

sand_mean <- aggregate_soilgrids(sand)
silt_mean <- aggregate_soilgrids(silt)
clay_mean <- aggregate_soilgrids(clay)
soc_mean <- aggregate_soilgrids(soc)

k <- k_factor(sand_mean, silt_mean, clay_mean, soc_mean)
k
#> class       : SpatRaster 
#> dimensions  : 10, 8, 1  (nrow, ncol, nlyr)
#> resolution  : 239.5216, 239.5216  (x, y)
#> extent      : 337313, 339229.2, 4814440, 4816835  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=utm +zone=38 +datum=WGS84 +units=m +no_defs 
#> source      : memory 
#> name        : K_williams1983 
#> min value   :     0.06520302 
#> max value   :     0.06743016